+++ options.go
@@ -108,6 +108,7 @@ var (
 		CreateKey:    "",
 		FullTypePath: false,
 		OmitNil:      false,
+		OmitEmpty:    false,
 		UseTags:      true,
 		KeyExact:     true,
 		NestEmbed:    false,
@@ -146,6 +147,11 @@ type Options struct {
 
 	// OmitNil skips the writing of nil values in an object.
 	OmitNil bool
+
+	// OmitEmpty skips the writing of empty string, slices, maps, and zero
+	// values although maps with all empty members will not be skipped on
+	// writing but will be with alt.Decompose and alter.
+	OmitEmpty bool
 
 	// InitSize is the initial buffer size.
 	InitSize int