+++ CHANGELOG.md
@@ -4,9 +4,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 The structure and content of this file follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.17.5] - unreleased
+## [1.17.5] - 2023-02-19
 ### Added
 - Added alt.Filter, a variation on alt.Match.
+- Added the OmitEmpty option to oj, sen, pretty, and alt packages.
+- Added the -o option for omit nil and empty to the oj command.
 
 ## [1.17.4] - 2023-02-02
 ### Fixed